+++ .rollup/local/exportsSveltePackage.js
@@ -2,7 +2,10 @@ import fs      from 'fs';
 import module  from 'module';
 
 const requireMod = module.createRequire(import.meta.url);
-const packageJSON = JSON.parse(fs.readFileSync(requireMod.resolve('@typhonjs-fvtt/svelte/package.json')).toString());
+const packageJSONPath = requireMod.resolve('@typhonjs-fvtt/svelte/package.json');
+const packageJSON = JSON.parse(fs.readFileSync(packageJSONPath).toString());
+
+const distPath = packageJSONPath.replace('package.json', '_dist');
 
 // `package.json` is skipped, but both Svelte application submodules need to be copied and paths converted separately.
 const s_EXCLUDE = ['./package.json', './application', './application/dialog', './application/legacy', './gsap/plugin/*',
@@ -10,6 +13,6 @@ const s_EXCLUDE = ['./package.json', './application', './application/dialog', '.
 
 // Formats exports from @typhonjs-fvtt/svelte package.json removing package.json reference and leading `.`.
 const exportsSveltePackage = Object.keys(packageJSON.exports).filter((entry) => !s_EXCLUDE.includes(entry)).map(
- (entry) => entry.substring(1));
+ (entry) => entry.substring(2));
 
-export { exportsSveltePackage };
+export { distPath, exportsSveltePackage };