+++ _dist/svelte/component/core/dialog/index.js
@@ -1,2 +1,0 @@
-export { default as DialogContent } from './DialogContent.svelte';
-export { default as DialogShell }   from './DialogShell.svelte';