+++ _dist/svelte/gsap/plugin/CustomEase.js
@@ -1,7 +1,0 @@
-import { gsap, gsapLoadPlugin } from '@typhonjs-fvtt/runtime/svelte/gsap';
-
-const CustomEase = await gsapLoadPlugin('CustomEase');
-
-gsap.registerPlugin(CustomEase);
-
-export { CustomEase };