+++ _dist/svelte/gsap/plugin/PixiPlugin.js
@@ -1,7 +1,0 @@
-import { gsap, gsapLoadPlugin } from '@typhonjs-fvtt/runtime/svelte/gsap';
-
-const PixiPlugin = await gsapLoadPlugin('PixiPlugin');
-
-gsap.registerPlugin(PixiPlugin);
-
-export { PixiPlugin };