+++ _dist/svelte/gsap/plugin/ScrollToPlugin.js
@@ -1,8 +1,0 @@
-import { gsap, gsapLoadPlugin } from '@typhonjs-fvtt/runtime/svelte/gsap';
-
-const ScrollToPlugin = await gsapLoadPlugin('ScrollToPlugin');
-
-gsap.registerPlugin(ScrollToPlugin);
-
-export { ScrollToPlugin };
-