+++ _dist/svelte/gsap/plugin/TextPlugin.js
@@ -1,7 +1,0 @@
-import { gsap, gsapLoadPlugin } from '@typhonjs-fvtt/runtime/svelte/gsap';
-
-const TextPlugin = await gsapLoadPlugin('TextPlugin');
-
-gsap.registerPlugin(TextPlugin);
-
-export { TextPlugin };