+++ _dist/svelte/gsap/plugin/bonus/GSDevTools.js
@@ -1,7 +1,0 @@
-import { gsap, gsapLoadPlugin } from '@typhonjs-fvtt/runtime/svelte/gsap';
-
-const GSDevTools = await gsapLoadPlugin('GSDevTools');
-
-gsap.registerPlugin(GSDevTools);
-
-export { GSDevTools };