+++ _dist/svelte/gsap/plugin/bonus/ScrambleTextPlugin.js
@@ -1,7 +1,0 @@
-import { gsap, gsapLoadPlugin } from '@typhonjs-fvtt/runtime/svelte/gsap';
-
-const ScrambleTextPlugin = await gsapLoadPlugin('ScrambleTextPlugin');
-
-gsap.registerPlugin(ScrambleTextPlugin);
-
-export { ScrambleTextPlugin };